safari-600.7-branch/LayoutTests/fast/canvas/canvas-color-clamping-expected.txt

Test that color values are properly clamped (R, G, B should be 0-255), alpha 0.0-1.0

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.shadowColor = 'rgba(0,0,0,0)'; ctx.shadowColor is 'rgba(0, 0, 0, 0)'
PASS ctx.shadowColor = 'rgb(0,0,0)'; ctx.shadowColor is '#000000'
PASS ctx.shadowColor = 'rgb(0,999,0)'; ctx.shadowColor is '#00ff00'
PASS ctx.shadowColor = 'rgb(0,999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999,0)'; ctx.shadowColor is '#00ff00'
PASS ctx.shadowColor = 'rgb(0,0,256)'; ctx.shadowColor is '#0000ff'
PASS ctx.shadowColor = 'rgb(999999999999999999999999,0,-9999999999999999999999999999)'; ctx.shadowColor is '#ff0000'
PASS ctx.shadowColor = 'rgba(9999999999999999999999999999999999999999999999999999999999999999999999999999999999,9,0,1)'; ctx.shadowColor is '#ff0900'
PASS ctx.shadowColor = 'rgba(9999999999999999999999999999999999999999999999999999999999999999999999999999999999,9,0,-99999999999999999999999999999999999999)'; ctx.shadowColor is 'rgba(255, 9, 0, 0)'
PASS ctx.shadowColor = 'rgba(7,9999999999999999999999999999999999999999999999999999999999999999999999999999999999,0,99999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999)'; ctx.shadowColor is '#07ff00'
PASS ctx.shadowColor = 'rgba(-7,9999999999999999999999999999999999999999999999999999999999999999999999999999999999,0,99999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999999)'; ctx.shadowColor is '#00ff00'
PASS ctx.shadowColor = 'rgba(0%,100%,0%,0.4)'; ctx.shadowColor is 'rgba(0, 255, 0, 0.4)'
PASS successfullyParsed is true

TEST COMPLETE
